+++ app/src/main/java/r48/maptools/UIMTAutotileRectangle.java
@@ -43,7 +43,7 @@ public class UIMTAutotileRectangle extends UIMTBase implements IMapViewCallbacks
     }
 
     @Override
-    public short shouldDrawAt(MapViewDrawContext.MouseStatus mouse, int tx, int ty, short there, int layer, int currentLayer) {
+    public int shouldDrawAt(MapViewDrawContext.MouseStatus mouse, int tx, int ty, int there, int layer, int currentLayer) {
         if (mouse != null) {
             int minX = Math.min(startX, mouse.x);
             int maxX = Math.max(startX, mouse.x);
