+++ app/src/main/java/r48/schema/specialized/MapPositionHelperSchemaElement.java
@@ -60,7 +60,7 @@ public class MapPositionHelperSchemaElement extends SchemaElement.Leaf {
         final UIMapView umv = new UIMapView(launcher.getApp(), mapGUM, 320, app.f.scaleGuess(192));
         umv.callbacks = new IMapViewCallbacks() {
             @Override
-            public short shouldDrawAt(MapViewDrawContext.MouseStatus mouse, int tx, int ty, short there, int layer, int currentLayer) {
+            public int shouldDrawAt(MapViewDrawContext.MouseStatus mouse, int tx, int ty, int there, int layer, int currentLayer) {
                 return there;
             }
